refactor: absolute imports, empty __init__.py, proper formatting
- Replace all relative imports with absolute imports
- Move facebook integration code from __init__.py to fi_facebook.py
- Empty all __init__.py files (facebook, admonster, owl_strategist)
- Add proper spacing between functions and classes for readability
