--- flexus_simple_bots/admonster/__init__.py
@@ -1,1 +1,0 @@
-# Ad Monster - LinkedIn Advertising Assistant
