--- flexus_simple_bots/owl_strategist/__init__.py
@@ -1,5 +1,0 @@
-# Owl Strategist bot package
-
-
-
-
